FIx bug where json_to_isatab.py writes files outside of project structure
There is a bug where whenever JsonToIsatabWriter is imported, the directory path tests/data/BII-I-1-generatedCombinedIsatab is always created from one level above the isa-api project root. This is because there is some residual test code embedded within the JsonToIsatabWriter class that always runs on import. This commit removes this so that these directories are not created. Test coverage is preserved by the appropriate test packages.

* Removed residual test code from end of JsonToIsatabWriter