Implement tearDown fixture
Without implementing a tearDown, the tests do not clean up after themselves since the tests create directories to write out ISA-Tab and ISA-JSON. This commit implements the tearDown to remove those generated directories after each test.

* Implement tearDown() fixture to remove generated directories and files
* Add to setup() to create temporary directories before each test is run
* Refactor test code to function correctly with setup() and tearDown() handling directory creation and cleanup

It is noted that the test_jsonToIsatab() test is dependent on IsatabtoJsonWriter() functioning correctly.