Refactor test for minimal unit test
Test was based on populating objects with a real data set, which is overkill. This commit removes all that dependency as really we only need to test minimal types and data structures in the objects.

* Remove real data set elements in test
* Implement test using as minimal information possible

Test needs to be sync'd with ISA v1 specification. Consequentially becomes a failing test until the v1 object model is updated.