Added better messaging
For some reason the bottom part of the function only logged errors and did not add them to the validator dictionary. They also did not indicate which filename the error was from, so tracking down where the problem is was difficult. This might just be an oversight since the warnings at the top of the function were just fine.