First attempt.
The problem with trying to do this is that some of the protocol types in the yaml file have "Sample Name" and "Extract Name" as the headers and this causes those headers to be added multiple times. It might be better to leave those empty. Also might be good to change headers to a list to get rid of the special case for "nucleic acid hybridization".